modals, spinners and other basic components added

+++ src/app/mileageLog/log-detail.component.ts
@@ -1,10 +1,13 @@
-import { Component, OnInit, OnDestroy } from '@angular/core';
+import { Component, Input, OnInit, OnDestroy } from '@angular/core';
 import { Router, ActivatedRoute, ROUTER_DIRECTIVES } from '@angular/router';
+import { Observable, Subscription } from 'rxjs/Rx';
 
 import { IMilageLog } from './mileage-log';
 import { MileageLogService } from './mileage-log.service';
 import { CapitalizePipe } from '../shared/pipes/capitalize.pipe';
-import { ToastService } from '../blocks/toast/toast.service';
+import { EntityService, ToastService, ModalService } from '../blocks/blocks';
+
+declare var componentHandler: any;
 
 @Component({
     moduleId: module.id,
@@ -15,14 +18,19 @@ import { ToastService } from '../blocks/toast/toast.service';
 })
 export class LogDetailComponent implements OnInit, OnDestroy {
     private sub: any;
-    log: IMilageLog;
+
+    @Input() log: IMilageLog;
+    editLog: IMilageLog = <IMilageLog>{};
 
     constructor(private router: Router,
         private route: ActivatedRoute,
         private _service: MileageLogService,
-        private _toast: ToastService) { }
+        private _toastService: ToastService,
+        private _modalService: ModalService,
+        private _entityService: EntityService) { }
 
     ngOnInit() {
+        componentHandler.upgradeDom();
         this.sub = this.router
             .routerState
             .queryParams
@@ -30,7 +38,11 @@ export class LogDetailComponent implements OnInit, OnDestroy {
                 console.info(params['id']);
                 let id = +params['id']; // (+) converts string 'id' to a number
                 this._service.getLog(id)
-                    .subscribe((log: IMilageLog) => this.log = log);
+                    .subscribe((log: IMilageLog) => this.log = log,
+                    (error) => {
+                        this._toastService.activate(`${error}`);
+                        return Observable.of();
+                    });
             });
     }
 
@@ -39,9 +51,21 @@ export class LogDetailComponent implements OnInit, OnDestroy {
     }
 
     deleteLog() {
-        let msg = 'Record successfully deleted';
-        this._service.deleteLog(this.log.id);
-        this._toast.activate(msg, 'Mileage Log');
-        window.history.back();
+        let msg = `Do you want to delete this record?`;
+        this._modalService.activate(msg).then(responseOK => {
+            if (responseOK) {
+                this.cancel(false);
+                this._service.deleteLog(this.log.id);
+                this._toastService.activate('Log successully deleted.', 'Mileage Log');
+                window.history.back();
+            }
+        });
+    }
+
+    cancel(showToast = true) {
+        this.editLog = this._entityService.clone(this.log);
+        if (showToast) {
+            this._toastService.activate(`Cancelled changes to entry`);
+        }
     }
 }
